Add command line option to not run barf turns
