+++ src/fights.ts
@@ -1,10 +1,7 @@
-import { canAdv } from "canadv.ash";
 import {
-  abort,
   adv1,
   availableAmount,
   booleanModifier,
-  chatPrivate,
   cliExecute,
   closetAmount,
   eat,
@@ -47,7 +44,6 @@ import {
   userConfirm,
   useSkill,
   visitUrl,
-  wait,
   weightAdjustment,
 } from "kolmafia";
 import {
@@ -92,7 +88,6 @@ import {
   Requirement,
   saleValue,
   setChoice,
-  sum,
 } from "./lib";
 import { freeFightMood, meatMood } from "./mood";
 import {
@@ -103,56 +98,7 @@ import {
 } from "./outfit";
 import { bathroomFinance } from "./potions";
 import { estimatedTurns, globalOptions, log } from "./globalvars";
-import { getString } from "libram/dist/property";
-
-function checkFax(): boolean {
-  if (!have($item`photocopied monster`)) cliExecute("fax receive");
-  if (getString("photocopyMonster") === "Knob Goblin Embezzler") return true;
-  cliExecute("fax send");
-  return false;
-}
-
-function faxEmbezzler(): void {
-  if (!get("_photocopyUsed")) {
-    if (checkFax()) return;
-    chatPrivate("cheesefax", "Knob Goblin Embezzler");
-    for (let i = 0; i < 3; i++) {
-      wait(10);
-      if (checkFax()) return;
-    }
-    abort("Failed to acquire photocopied Knob Goblin Embezzler.");
-  }
-}
-
-type EmbezzlerFightOptions = {
-  location?: Location;
-  macro?: Macro;
-};
-
-class EmbezzlerFight {
-  available: () => boolean;
-  potential: () => number;
-  run: (options: EmbezzlerFightOptions) => void;
-  requirements: Requirement[];
-  draggable: boolean;
-  name: string;
-
-  constructor(
-    name: string,
-    available: () => boolean,
-    potential: () => number,
-    run: (options: EmbezzlerFightOptions) => void,
-    requirements: Requirement[] = [],
-    draggable = false
-  ) {
-    this.name = name;
-    this.available = available;
-    this.potential = potential;
-    this.run = run;
-    this.requirements = requirements;
-    this.draggable = draggable;
-  }
-}
+import { embezzlerCount, EmbezzlerFight, embezzlerMacro, embezzlerSources } from "./embezzlers";
 
 const firstChainMacro = () =>
   Macro.if_(
@@ -192,199 +138,6 @@ const secondChainMacro = () =>
         .trySkill("Lecture on Relativity")
     ).meatKill()
   ).abort();
-
-const embezzlerMacro = () =>
-  Macro.if_(
-    "monstername Knob Goblin Embezzler",
-    Macro.if_("snarfblat 186", Macro.tryCopier($item`pulled green taffy`))
-      .trySkill("Wink At")
-      .trySkill("Fire a badly romantic arrow")
-      .externalIf(
-        get("_sourceTerminalDigitizeMonster") !== $monster`Knob Goblin Embezzler`,
-        Macro.tryCopier($skill`Digitize`)
-      )
-      .tryCopier($item`Spooky Putty sheet`)
-      .tryCopier($item`Rain-Doh black box`)
-      .tryCopier($item`4-d camera`)
-      .tryCopier($item`unfinished ice sculpture`)
-      .meatKill()
-  ).abort();
-
-const embezzlerSources = [
-  new EmbezzlerFight(
-    "Digitize",
-    () =>
-      get("_sourceTerminalDigitizeMonster") === $monster`Knob Goblin Embezzler` &&
-      getCounters("Digitize Monster", 0, 0).trim() !== "",
-    () => (SourceTerminal.have() && get("_sourceTerminalDigitizeUses") === 0 ? 1 : 0),
-    (options: EmbezzlerFightOptions) => {
-      adv1(options.location || $location`Noob Cave`);
-    },
-    [],
-    true
-  ),
-  new EmbezzlerFight(
-    "Backup",
-    () =>
-      get("lastCopyableMonster") === $monster`Knob Goblin Embezzler` &&
-      have($item`backup camera`) &&
-      get<number>("_backUpUses") < 11,
-    () => (have($item`backup camera`) ? 11 - get<number>("_backUpUses") : 0),
-    (options: EmbezzlerFightOptions) => {
-      const realLocation =
-        options.location && options.location.combatPercent >= 100
-          ? options.location
-          : $location`Noob Cave`;
-      adventureMacro(
-        realLocation,
-        Macro.if_(
-          "!monstername Knob Goblin Embezzler",
-          Macro.skill("Back-Up to Your Last Enemy")
-        ).step(options.macro || embezzlerMacro())
-      );
-    },
-    [
-      new Requirement([], {
-        forceEquip: $items`backup camera`,
-        bonusEquip: new Map([[$item`backup camera`, 5000]]),
-      }),
-    ],
-    true
-  ),
-  new EmbezzlerFight(
-    "Fax",
-    () => have($item`Clan VIP Lounge key`) && !get("_photocopyUsed"),
-    () => (have($item`Clan VIP Lounge key`) && !get("_photocopyUsed") ? 1 : 0),
-    () => {
-      faxEmbezzler();
-      use($item`photocopied monster`);
-    }
-  ),
-  new EmbezzlerFight(
-    "Pillkeeper Semirare",
-    () =>
-      have($item`Eight Days a Week Pill Keeper`) &&
-      canAdv($location`Cobb's Knob Treasury`, true) &&
-      !get("_freePillKeeperUsed"),
-    () =>
-      have($item`Eight Days a Week Pill Keeper`) &&
-      canAdv($location`Cobb's Knob Treasury`, true) &&
-      !get("_freePillKeeperUsed")
-        ? 1
-        : 0,
-    () => {
-      cliExecute("pillkeeper semirare");
-      adv1($location`Cobb's Knob Treasury`);
-    }
-  ),
-  new EmbezzlerFight(
-    "Chateau Painting",
-    () =>
-      ChateauMantegna.have() &&
-      !ChateauMantegna.paintingFought() &&
-      ChateauMantegna.paintingMonster() === $monster`Knob Goblin Embezzler`,
-    () =>
-      ChateauMantegna.have() &&
-      !ChateauMantegna.paintingFought() &&
-      ChateauMantegna.paintingMonster() === $monster`Knob Goblin Embezzler`
-        ? 1
-        : 0,
-    () => ChateauMantegna.fightPainting()
-  ),
-  new EmbezzlerFight(
-    "Spooky Putty & Rain-Doh",
-    () =>
-      (have($item`Spooky Putty monster`) &&
-        get("spookyPuttyMonster") === $monster`Knob Goblin Embezzler`) ||
-      (have($item`Rain-Doh box full of monster`) &&
-        get("rainDohMonster") === $monster`Knob Goblin Embezzler`),
-    () => {
-      if (
-        (have($item`Spooky Putty sheet`) || have($item`Spooky Putty monster`)) &&
-        (have($item`Rain-Doh black box`) || have($item`Rain-Doh box full of monster`))
-      ) {
-        return (
-          6 -
-          get("spookyPuttyCopiesMade") -
-          get("_raindohCopiesMade") +
-          itemAmount($item`Spooky Putty monster`) +
-          itemAmount($item`Rain-Doh box full of monster`)
-        );
-      } else if (have($item`Spooky Putty sheet`) || have($item`Spooky Putty monster`)) {
-        return 5 - get("spookyPuttyCopiesMade") + itemAmount($item`Spooky Putty monster`);
-      } else if (have($item`Rain-Doh black box`) || have($item`Rain-Doh box full of monster`)) {
-        return 5 - get("_raindohCopiesMade") + itemAmount($item`Rain-Doh box full of monster`);
-      }
-      return 0;
-    },
-    () => {
-      if (have($item`Spooky Putty monster`)) return use($item`Spooky Putty monster`);
-      return use($item`Rain-Doh box full of monster`);
-    }
-  ),
-  new EmbezzlerFight(
-    "4-d Camera",
-    () =>
-      have($item`shaking 4-d camera`) &&
-      get("cameraMonster") === $monster`Knob Goblin Embezzler` &&
-      !get("_cameraUsed"),
-    () =>
-      have($item`shaking 4-d camera`) &&
-      get("cameraMonster") === $monster`Knob Goblin Embezzler` &&
-      !get("_cameraUsed")
-        ? 1
-        : 0,
-    () => use($item`shaking 4-d camera`)
-  ),
-  new EmbezzlerFight(
-    "Ice Sculpture",
-    () =>
-      have($item`ice sculpture`) &&
-      get("iceSculptureMonster") === $monster`Knob Goblin Embezzler` &&
-      !get("_iceSculptureUsed"),
-    () =>
-      have($item`ice sculpture`) &&
-      get("iceSculptureMonster") === $monster`Knob Goblin Embezzler` &&
-      !get("_iceSculptureUsed")
-        ? 1
-        : 0,
-    () => use($item`ice sculpture`)
-  ),
-  new EmbezzlerFight(
-    "Green Taffy",
-    () =>
-      have($item`envyfish egg`) &&
-      get("envyfishMonster") === $monster`Knob Goblin Embezzler` &&
-      !get("_envyfishEggUsed"),
-    () =>
-      have($item`envyfish egg`) &&
-      get("envyfishMonster") === $monster`Knob Goblin Embezzler` &&
-      !get("_envyfishEggUsed")
-        ? 1
-        : 0,
-    () => use($item`envyfish egg`)
-  ),
-  new EmbezzlerFight(
-    "Professor MeatChain",
-    () => false,
-    () => (have($familiar`Pocket Professor`) && !get<boolean>("_garbo_meatChain", false) ? 10 : 0),
-    () => {
-      return;
-    }
-  ),
-  new EmbezzlerFight(
-    "Professor WeightChain",
-    () => false,
-    () => (have($familiar`Pocket Professor`) && !get<boolean>("_garbo_weightChain", false) ? 5 : 0),
-    () => {
-      return;
-    }
-  ),
-];
-
-export function embezzlerCount(): number {
-  return sum(embezzlerSources.map((source) => source.potential()));
-}
 
 function embezzlerSetup() {
   meatMood(true, true).execute(estimatedTurns());
