+++ src/index.ts
@@ -309,6 +309,9 @@ export function main(argString = ""): void {
     if (arg.match(/ascend/)) {
       globalOptions.ascending = true;
     }
+    if (arg.match(/nobarf/)) {
+      globalOptions.noBarf = true;
+    }
   }
   const gardens = $items`packet of pumpkin seeds, Peppermint Pip Packet, packet of dragon's teeth, packet of beer seeds, packet of winter seeds, packet of thanksgarden seeds, packet of tall grass seeds, packet of mushroom spores`;
   const startingGarden = gardens.find((garden) =>
@@ -392,28 +395,30 @@ export function main(argString = ""): void {
         // 0. diet stuff.
         runDiet();
 
-        // 1. get a ticket
-        ensureBarfAccess();
-
-        // 2. make an outfit (amulet coin, pantogram, etc), misc other stuff (VYKEA, songboom, robortender drinks)
+        // 1. make an outfit (amulet coin, pantogram, etc), misc other stuff (VYKEA, songboom, robortender drinks)
         dailySetup();
 
         setDefaultMaximizeOptions({
           preventEquip: $items`broken champagne bottle, Spooky Putty snake, Spooky Putty mitre, Spooky Putty leotard, Spooky Putty ball, papier-mitre`,
         });
 
-        // 4. do some embezzler stuff
+        // 2. do some embezzler stuff
         freeFights();
         dailyFights();
 
-        // 5. burn turns at barf
-        try {
-          while (canContinue()) {
-            barfTurn();
+        if (!globalOptions.noBarf) {
+          // 3. get a ticket
+          ensureBarfAccess();
+
+          // 4. burn turns at barf
+          try {
+            while (canContinue()) {
+              barfTurn();
+            }
+          } finally {
+            setAutoAttack(0);
           }
-        } finally {
-          setAutoAttack(0);
-        }
+        } else setAutoAttack(0);
       });
     });
   } finally {
